fixed cstyle casts

+++ Chess/ChessPeek/ChessPeek/src/ChessPeek/Utilities.cpp
@@ -16,7 +16,9 @@ XY convertToXY(BitmapXY const& bitmapXY) { return bitmapXY; }
 BitmapXY convertToBitmapXY(XY const& xy) { return xy; }
 
 double calculateColorIntensityDecimal(uint32_t const color) {
-    return (((double)extractRed(color) + extractGreen(color) + extractBlue(color)) / 0xFF) / 3;
+    return (static_cast<double>(static_cast<uint32_t>(extractRed(color)) + extractGreen(color) + extractBlue(color)) /
+            0xFF) /
+           3;
 }
 
 uint8_t extractRed(uint32_t const color) { return (AlbaBitManipulation<uint32_t>::getByteAt<2>(color)); }
